// src/api/mangadex/models/manga.ts
/**
 * Generated by orval v6.31.0 🍺
 * Do not edit manually.
 * MangaDex API
 * MangaDex is an ad-free manga reader offering high-quality images!

This document details our API as it is right now. It is in no way a promise to never change it, although we will endeavour to publicly notify any major change.

# Acceptable use policy

Usage of our services implies acceptance of the following:
- You **MUST** credit us
- You **MUST** credit scanlation groups if you offer the ability to read chapters
- You **CANNOT** run ads or paid services on your website and/or apps

These may change at any time for any and no reason and it is up to you check for updates from time to time.

# Security issues

If you believe you found a security issue in our API, please check our [security.txt](/security.txt) to get in touch privately.

 * OpenAPI spec version: 5.10.2
 */
import type { MangaAttributes } from './mangaAttributes';
import type { Relationship } from './relationship';
import type { MangaType } from './mangaType';

export interface Manga {
  attributes?: MangaAttributes;
  id?: string;
  relationships?: Relationship[];
  type?: MangaType;
}